# Improves the performance of VariableBaseMSM when scalars are small