--- a/ec/src/hashing/curve_maps/elligator2.rs
+++ b/ec/src/hashing/curve_maps/elligator2.rs
@@ -279,7 +279,7 @@ mod test {
     fn map_field_to_curve_elligator2() {
         Elligator2Map::<TestElligator2MapToCurveConfig>::check_parameters().unwrap();
 
-        let mut map_range: Vec<Affine<TestElligator2MapToCurveConfig>> = vec![];
+        let mut map_range: Vec<Affine<TestElligator2MapToCurveConfig>> = Vec::new();
         // We are mapping all elements of the field to the curve, verifying that
         // map is not constant on that set.
         for current_field_element in 0..101 {