--- a/ff/src/const_helpers.rs
+++ b/ff/src/const_helpers.rs
@@ -358,7 +358,7 @@ mod tests {
     fn test_ser_buffer_write_and_read() {
         type Ser = SerBuffer<2>;
         let buf = Ser::zeroed();
-        let mut data = vec![];
+        let mut data = ark_std::vec::Vec::new();
         buf.write_up_to(&mut data, 16)
             .expect("Failed to write buffer");