<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v                               ✰  Thanks for creating a PR! ✰
v    Before hitting that submit button please review the checkboxes.
v    If a checkbox is n/a - please still include it but + a little note why
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

## Description

<!-- Add a description of the changes that this PR introduces and the files that
are the most critical to review.
-->

Greatly improves performance when the size of the scalars is known to be small.
Inspired by [`jolt`'s MSMs](https://github.com/a16z/jolt/blob/e8c245607a4ab881f40713bae7cc4119dc117ae4/jolt-core/src/msm/mod.rs), except the version in this PR automatically partitions scalars and bases according to scalar size, and uses the best MSM for each partition.

```
MSM/MSM-random for BN254::G1
                        time:   [594.29 ms 595.72 ms 597.40 ms]
                        change: [+0.5766% +1.1131% +1.5745%] (p = 0.00 < 0.05)
                        Change within noise threshold.
MSM/MSM-bool for BN254::G1
                        time:   [15.240 ms 15.341 ms 15.427 ms]
                        change: [−86.632% −86.501% −86.377%] (p = 0.00 < 0.05)
                        Performance has improved.
MSM/MSM-u8 for BN254::G1
                        time:   [29.244 ms 29.361 ms 29.497 ms]
                        change: [−84.803% −84.722% −84.645%] (p = 0.00 < 0.05)
                        Performance has improved.
MSM/MSM-i8 for BN254::G1
                        time:   [32.387 ms 33.462 ms 34.364 ms]
                        change: [−89.208% −88.956% −88.652%] (p = 0.00 < 0.05)
                        Performance has improved.
MSM/MSM-u16 for BN254::G1
                        time:   [51.980 ms 52.116 ms 52.218 ms]
                        change: [−76.408% −76.274% −76.141%] (p = 0.00 < 0.05)
                        Performance has improved.
MSM/MSM-i16 for BN254::G1
                        time:   [54.833 ms 55.824 ms 56.835 ms]
                        change: [−82.000% −81.774% −81.554%] (p = 0.00 < 0.05)
                        Performance has improved.
MSM/MSM-u32 for BN254::G1
                        time:   [79.570 ms 79.736 ms 79.889 ms]
                        change: [−64.927% −64.787% −64.639%] (p = 0.00 < 0.05)
                        Performance has improved.
MSM/MSM-i32 for BN254::G1
                        time:   [86.232 ms 86.877 ms 87.974 ms]
                        change: [−71.781% −71.583% −71.367%] (p = 0.00 < 0.05)
                        Performance has improved.
MSM/MSM-u64 for BN254::G1
                        time:   [130.77 ms 130.95 ms 131.18 ms]
                        change: [−43.628% −43.224% −42.792%] (p = 0.00 < 0.05)
                        Performance has improved.
MSM/MSM-i64 for BN254::G1
                        time:   [144.81 ms 145.62 ms 146.09 ms]
                        change: [−61.201% −59.247% −57.091%] (p = 0.00 < 0.05)
                        Performance has improved.
MSM/MSM-mixed for BN254::G1
                        time:   [135.97 ms 136.29 ms 136.59 ms]
                        change: [−58.709% −58.510% −58.313%] (p = 0.00 < 0.05)
                        Performance has improved.
```

```
MSM/MSM-random for BN254::G2
                        time:   [1.6892 s 1.6924 s 1.6960 s]
                        change: [+0.3237% +0.5644% +0.8446%] (p = 0.00 < 0.05)
                        Change within noise threshold.
MSM/MSM-bool for BN254::G2
                        time:   [39.202 ms 39.288 ms 39.365 ms]
                        change: [−86.115% −86.063% −86.009%] (p = 0.00 < 0.05)
                        Performance has improved.
MSM/MSM-u8 for BN254::G2
                        time:   [76.824 ms 76.965 ms 77.132 ms]
                        change: [−85.368% −85.325% −85.270%] (p = 0.00 < 0.05)
                        Performance has improved.
MSM/MSM-i8 for BN254::G2
                        time:   [81.739 ms 82.361 ms 83.092 ms]
                        change: [−90.137% −90.043% −89.954%] (p = 0.00 < 0.05)
                        Performance has improved.
MSM/MSM-u16 for BN254::G2
                        time:   [143.33 ms 143.95 ms 144.34 ms]
                        change: [−75.856% −75.567% −75.279%] (p = 0.00 < 0.05)
                        Performance has improved.
MSM/MSM-i16 for BN254::G2
                        time:   [149.28 ms 150.47 ms 151.19 ms]
                        change: [−82.237% −82.099% −81.963%] (p = 0.00 < 0.05)
                        Performance has improved.
MSM/MSM-u32 for BN254::G2
                        time:   [224.01 ms 224.53 ms 225.08 ms]
                        change: [−63.798% −63.557% −63.315%] (p = 0.00 < 0.05)
                        Performance has improved.
MSM/MSM-i32 for BN254::G2
                        time:   [241.36 ms 242.28 ms 243.26 ms]
                        change: [−71.275% −71.153% −71.029%] (p = 0.00 < 0.05)
                        Performance has improved.
MSM/MSM-u64 for BN254::G2
                        time:   [374.23 ms 374.84 ms 375.46 ms]
                        change: [−38.459% −38.175% −37.881%] (p = 0.00 < 0.05)
                        Performance has improved.
MSM/MSM-i64 for BN254::G2
                        time:   [409.37 ms 410.56 ms 411.59 ms]
                        change: [−63.278% −60.759% −57.586%] (p = 0.00 < 0.05)
                        Performance has improved.
MSM/MSM-mixed for BN254::G2
                        time:   [366.64 ms 369.24 ms 373.68 ms]
                        change: [−58.802% −58.470% −57.970%] (p = 0.00 < 0.05)
                        Performance has improved.
  
```

---

Before we can merge this PR, please make sure that all the following items have been
checked off. If any of the checklist items are not applicable, please leave them but
write a little note why.

- [ ] Targeted PR against correct branch (master)
- [ ] Linked to GitHub issue with discussion and accepted design OR have an explanation in the PR that describes this work.
- [ ] Wrote unit tests
- [ ] Updated relevant documentation in the code
- [ ] Added a relevant changelog entry to the `Pending` section in `CHANGELOG.md`
- [ ] Re-reviewed `Files changed` in the GitHub PR explorer