@Pratyush 2025-06-18
Just pushed a new commit that improves parallelization for even the standard wNAF MSM. See below:

```
MSM/MSM-random for BN254::G1
                        time:   [515.81 ms 516.54 ms 517.38 ms]
                        change: [−14.531% −14.091% −13.685%] (p = 0.00 < 0.05)
                        Performance has improved.
                        ```